Added all queries and steps done.
Updated all queries and steps undergone for acquirers in Harvard DataVerse and why the codes of acquirers have to be modified in Knowledge Space NBER considering all gvkeys and updated all output files and carried further steps in Harvard with these files. Finally left with final two files of both acquirer and target. Also added problem in the conversion of asgnum to gvkey and steps trying to resolve this issue which hasn't yet been solved.